pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler==3.10.4",
    "flask-cors==4.0.0",
    "flask-limiter==3.5.0",
    "flask-sqlalchemy==3.1.1",
    "flask==2.3.3",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "sqlalchemy==2.0.23",
    "flask-jwt-extended==4.5.3",
]
